requirements.txt
streamlit
requests
uvicorn
pymupdf
fastapi
pydantic
langchain
langchain-google-genai
openai
chromadb
python-dotenv
langchain_chroma
python-multipart